# Some perf tweaks to the tokenizer.