@Ms2ger 2017-01-18
> The stylesheet is a mix of MozReview, The Guardian, and Facebook, so I can remove it if there's any concern about having it in-repo (we can find others also).

Yes, please remove.