@emilio 2017-01-18
> Yes, please remove.

Done