@SimonSapin 2017-01-20
Reviewed 1 of 1 files at r1, 7 of 7 files at r2, 3 of 3 files at r3, 1 of 1 files at r4, 1 of 1 files at r5, 1 of 1 files at r7.
Review status: all files reviewed at latest revision, 13 unresolved discussions.

---

*[.travis.yml, line 13 at r6](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KariN9U9zn6jxTS-Sm7:-KariN9U9zn6jxTS-Sm8:bd6q46l) ([raw file](https://github.com/servo/rust-cssparser/blob/a579c022255a71e30e451aff037d161577d8f32d/.travis.yml#L13)):*
> ```YAML
>   - cargo test --features heapsize
>   - cargo test --features dummy_match_byte
>   - cargo bench --features bench
> ```

Why is it useful to run benchmarks on CI? They take time, and the CPU speed of VMs can vary a lot (e.g if another Travis job is running on the same CPU at the same time).

---

*[build.rs, line 1 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KargD9E-46brCqTK9lT:-KargD9E-46brCqTK9lU:buvxg0x) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/build.rs#L1)):*
> ```Rust
> 
> ```

Needs a copyright header.

---

*[src/tokenizer.rs, line 408 at r1](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KarfF5_RtRav2jbIoHi:-KarfF5_RtRav2jbIoHj:b80wyss) ([raw file](https://github.com/servo/rust-cssparser/blob/afcf259b3ae7f34e913e9b26e7ca28b06c2271c9/src/tokenizer.rs#L408)):*
> ```Rust
>         return None
>     }
>     let c = tokenizer.next_byte_unchecked();
> ```

Please rename `c` to `b` or `byte`.

---

*[src/tokenizer.rs, line 556 at r1](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-Karexgv6SVBGUhJOh3V:-KarexgwPJ4hEX6Kfass:b-uabagw) ([raw file](https://github.com/servo/rust-cssparser/blob/afcf259b3ae7f34e913e9b26e7ca28b06c2271c9/src/tokenizer.rs#L556)):*
> ```Rust
>                 consume_ident_like(tokenizer)
>             } else {
>                 let ret = Delim(tokenizer.next_char());
> ```

Since we’re already excluding non-ASCII, `c as u8` can be used instead of `next_char()`. (And `ret` becomes unnecessary.)

---

*[src/tokenizer.rs, line 409 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KavEsj7_mt0oHTyJzsy:-KavEsj8ry9U9DN1sHsV:b-484s0d) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/tokenizer.rs#L409)):*
> ```Rust
>     }
>     let c = tokenizer.next_byte_unchecked();
>     let token = match_byte! { c,
> ```

Have you checked that `match_byte!` generates better code than `match`?

---

*[src/macros/match_byte.rs, line 1 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KargMI0fLUkBJj3ehGy:-KargMI1C7ZcwyoyhUly:buvxg0x) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/match_byte.rs#L1)):*
> ```Rust
> 
> ```

Needs a copyright header.

---

*[src/macros/match_byte.rs, line 53 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-Kav8sQTyHKkYhXf59W0:-Kav8sQUiGSG8Lj-sZrm:b3xvwat) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/match_byte.rs#L53)):*
> ```Rust
> }
> 
> fn parse_match_bytes_macro(tts: Vec<syn::TokenTree>) -> (Vec<syn::TokenTree>, Vec<u8>, Vec<Case>, Option<syn::Ident>) {
> ```

This could use some docs. What is the expected "shape" of the input tokens (maybe give an example)? What are the respective components of the return value?

---

*[src/macros/match_byte.rs, line 82 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KavDS5CPQAvj470SqLQ:-KavDS5CPQAvj470SqLR:bvt8zqu) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/match_byte.rs#L82)):*
> ```Rust
>     let mut tts = tts.peekable();
>     let mut case_id: u8 = 1;
>     let mut binding = None;
> ```

The parser allows multiple bindings but only the last one will work. But since this macro is not public and only used with a known set of inputs maybe that’s ok.

---

*[src/macros/match_byte.rs, line 108 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KavDA6w79aLYZqTyeus:-KavDA6xslPFVSUWboSX:b3psh4t) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/match_byte.rs#L108)):*
> ```Rust
> /// The `binding` parameter is the identifier that is used by the wildcard
> /// pattern.
> fn parse_case(tts: &mut iter::Peekable<vec::IntoIter<syn::TokenTree>>,
> ```

This accepts some inputs that it shouldn’t (the empty input, an ident followed by a byte pattern, …) but since this macro is not public and only used with a known set of inputs maybe that’s ok.

---

*[src/macros/match_byte.rs, line 182 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KavDbVpCzFrOijFnRXx:-KavDbVpCzFrOijFnRXy:b5bmwcx) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/match_byte.rs#L182)):*
> ```Rust
> 
>     assert!(!to_be_matched.is_empty());
>     assert_eq!(table.len(), 256);
> ```

Replace `Vec<u8>` with `[u8; 256]` to make this assert unnecessary?

---

*[src/macros/mod.rs, line 1 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KargVfncrGiNMb0CST2:-KargVfoGBZbDhU9Bc4n:buvxg0x) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/mod.rs#L1)):*
> ```Rust
> 
> ```

Needs a copyright header.

---

*[src/macros/visit.rs, line 1 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-Karga8ffglm6KdhooGF:-Karga8ffglm6KdhooGG:b-l1p9yj) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/visit.rs#L1)):*
> ```Rust
> // Copyright 2016 The html5ever Project Developers. See the
> ```

I believe I wrote all of this code, and I agree to re-licensing it under the MPL 2. Please change this license header to the same as the rest of cssparser.

---

*[src/macros/visit.rs, line 14 at r2](https://reviewable.io:443/reviews/servo/rust-cssparser/112#-KargiUcInknqkBRrern:-KargiUcInknqkBRrero:b-1lf94a) ([raw file](https://github.com/servo/rust-cssparser/blob/9c955423498898ac7a7796dbf03554cee96fd5fa/src/macros/visit.rs#L14)):*
> ```Rust
> use syn;
> 
> pub trait Visitor {
> ```

In html5ever I initially did that but then switched to only parsing token trees and avoiding the AST entirely, so that a visitor trait is not needed. Was there a reason not to do that here?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/servo/rust-cssparser/112)*
<!-- Sent from Reviewable.io -->